+++ CHANGELOG.md
@@ -5,23 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.3]
 
 ## [0.0.2]
 
 ### Added
 
 - Get-IpNetworkBase
-- Documentation for new function
+- Documentation for new functions
+- Convert-IpToMaskLen
+- Get-IpNetworkEndIp
+- Get-IpNetworkStartIp
+- Test-IpRangeIsSubnet
+- Test-ValidMask
 
 ### Updated
 
 - IpNetwork Class instantiator uses Get-IpNetworkBase to set the IpAddress of the resultant object
 - about_Ip.Tools discusses use of Class
 - Pester Tests now load module using **Using**
+- jscpd default configuration with threshold
 
 ### Fixed
 
+- Various documentation example inconsistencies
 - Not using the Network Base address when defining an IpNetwork caused some Test-IpInNetwork tests to fail
 - Linting issues from first run(s) of MegaLinter on this repo